Refactor ipsum service and word
